apply clippy

+++ zaino-testutils/src/lib.rs
@@ -463,7 +463,7 @@ mod launch_testmanager {
             .unwrap();
             assert_eq!(
                 2,
-                u32::from(test_manager.local_net.get_chain_height().await)
+                (test_manager.local_net.get_chain_height().await)
             );
             test_manager.close().await;
         }
@@ -486,12 +486,12 @@ mod launch_testmanager {
             .unwrap();
             assert_eq!(
                 2,
-                u32::from(test_manager.local_net.get_chain_height().await)
+                (test_manager.local_net.get_chain_height().await)
             );
             test_manager.local_net.generate_blocks(1).await.unwrap();
             assert_eq!(
                 3,
-                u32::from(test_manager.local_net.get_chain_height().await)
+                (test_manager.local_net.get_chain_height().await)
             );
             test_manager.close().await;
         }
@@ -515,7 +515,7 @@ mod launch_testmanager {
             .unwrap();
             assert_eq!(
                 10,
-                u32::from(test_manager.local_net.get_chain_height().await)
+                (test_manager.local_net.get_chain_height().await)
             );
             test_manager.close().await;
         }
@@ -645,7 +645,7 @@ mod launch_testmanager {
                     .unwrap();
                 assert_eq!(
                     2,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.close().await;
             }
@@ -669,12 +669,12 @@ mod launch_testmanager {
                     .unwrap();
                 assert_eq!(
                     2,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.local_net.generate_blocks(1).await.unwrap();
                 assert_eq!(
                     3,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.close().await;
             }
@@ -699,7 +699,7 @@ mod launch_testmanager {
                     .unwrap();
                 assert_eq!(
                     52,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.close().await;
             }
@@ -965,7 +965,7 @@ mod launch_testmanager {
                     .unwrap();
                 assert_eq!(
                     2,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.close().await;
             }
@@ -989,12 +989,12 @@ mod launch_testmanager {
                     .unwrap();
                 assert_eq!(
                     2,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.local_net.generate_blocks_with_delay(1).await;
                 assert_eq!(
                     3,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.close().await;
             }
@@ -1019,7 +1019,7 @@ mod launch_testmanager {
                     .unwrap();
                 assert_eq!(
                     52,
-                    u32::from(test_manager.local_net.get_chain_height().await)
+                    (test_manager.local_net.get_chain_height().await)
                 );
                 test_manager.close().await;
             }
